refactor: improving the folder structure of features layers
