+++ apps/front/src/features/food-slot-machine/FoodSlotMachine.tsx
