+++ apps/front/src/features/food-slot-machine/constants.ts
