+++ apps/front/src/features/food-slot-machine/index.tsx
@@ -1,0 +1,1 @@
+export { FoodSlotMachine } from "./FoodSlotMachine";
