+++ apps/front/src/features/glitch-text/GlitchText.tsx
