+++ apps/front/src/features/glitch-text/index.tsx
@@ -1,0 +1,1 @@
+export { GlitchText } from "./GlitchText";
