+++ apps/front/src/features/interaction-card/InteractionCard.tsx
