+++ apps/front/src/features/interaction-card/hooks.ts
