+++ apps/front/src/features/interaction-card/index.tsx
@@ -1,0 +1,1 @@
+export { InteractionCard } from "./InteractionCard";
