+++ apps/front/src/features/interaction-card/types.ts
