+++ apps/front/src/features/interaction-card/utils.ts
